refactor(models): partition UI and GameToUI event libraries (#4334)
Split ui_action_events and game_to_ui_events into domain first-class
libraries with thin barrels, matching the session_command partition pattern.
Public types and barrel exports stay stable; sealed bases become abstract
to allow cross-file subclasses without part directives.
