+++ packages/colonizethis_models/lib/src/app_events.dart
@@ -13,10 +13,11 @@
 // [OpenPanelEvent] remains for legacy string-id panels until migrated.
 //
 // Concrete event types live under `app_events/` as first-class libraries
-// (Refs #4068 Slice C): `ui_action_events.dart` (UIActionEvent),
-// `session_command_events.dart` barrel (SessionCommandEvent + domain partitions),
-// `ui_system_events.dart` (UISystemEvent), and `game_to_ui_events.dart`
-// (GameToUIEvent + app-prefixed GameEvent mirrors).
+// (Refs #4068 Slice C; wave-3 partitions Refs #4334): `ui_action_events.dart`
+// barrel (UIActionEvent + domain partitions), `session_command_events.dart`
+// barrel (SessionCommandEvent + domain partitions), `ui_system_events.dart`
+// (UISystemEvent), and `game_to_ui_events.dart` barrel (GameToUIEvent +
+// app-prefixed GameEvent mirror partitions).
 
 
 export 'ai_events.dart' show DialogueEvent, PortraitMoodEvent;
